Adding in container images, helm charts and GH actions workflows that closely mirror those used for conch (so that we have complete consistency)
Currently I am building images / helm charts for just op-portal and op-bridge,
as these are the ones to be deployed on k8s in AWS. The other agent executables
will likely be systemd daemons? Also I am not yet building the python module
which would need to be uploaded to, e.g. pip?

This is likely to be very broken