Added a disconnect function so that a connection can be disconnected if the keepalive fails to send.
Also added a watchdog command - a watchdog checks the status of
the connection periodically. If there has been no activity for
the last 5 minutes, then the connection is disconnected and
remade.

This should prevent situations where the connection freezes
because one side thinks (wrongly) that it is still open.

Also added a guard that prevents a keepalive (or watchdog)
flood - only a single watchdog or keepalive can be in flight
per connection.